Rule based in Btch File, Bugs fixed.

+++ Name and Address Parser/submission_form_parser.py
@@ -140,7 +140,7 @@ class submission_form:
             print(f"Token Pattern: {pattern}")
             print(f"Comment: {Comment}")
     
-            msg.showinfo("Info", "Address Added to Validation DataBase!")
+            msg.showinfo("Info", "Address Added to Validation DataBase And Moved to Knowledge Base!")
             print(f"Approved By: {approval} at {today}")
             components = form_frame.winfo_children()
             scrollbar.destroy()
@@ -166,6 +166,7 @@ class submission_form:
                 approval_form_instance.Browse_File(df,True,form_frame,canvas,table_frame,label1,label2,label3,tab4)
                 scrollbar.destroy()
                 table_inner_frame.destroy()
+            
             
         elif toggle_state.get() == "No":
             msg.showinfo("Info", "Address is not Approved!")#\nPlease Select a New Exception File")
@@ -215,26 +216,27 @@ class submission_form:
         
         for data in table_data:
             print(data)
+        if toggle_state.get() == "Yes":
+            i=1
+            Def_Dict={}
+            for n in table_data:
+                if n[2] in Def_Dict.keys():
+                    print(Def_Dict[n[2]])
+                    Val=Def_Dict[n[2]]
+                    Val.append(i)
+                    Def_Dict[n[2]]=Val
+                else:
+                    Def_Dict[n[2]]=[i]
+                i+=1
+    
+            print(Def_Dict)
+            with open('KB_Kashif.json', 'r+', encoding='utf-8') as f:
+                data = json.load(f)          
+                data[pattern] = Def_Dict# <--- add `id` value.
+                f.seek(0)        # <--- should reset file position to the beginning.
+                json.dump(data, f)
+                f.truncate()# remove remaining part
+            return  
         
-        i=1
-        Def_Dict={}
-        for n in table_data:
-            if n[2] in Def_Dict.keys():
-                print(Def_Dict[n[2]])
-                Val=Def_Dict[n[2]]
-                Val.append(i)
-                Def_Dict[n[2]]=Val
-            else:
-                Def_Dict[n[2]]=[i]
-            i+=1
-
-        print(Def_Dict)
-        with open('KB_Kashif.json', 'r+', encoding='utf-8') as f:
-            data = json.load(f)          
-            data[pattern] = Def_Dict# <--- add `id` value.
-            f.seek(0)        # <--- should reset file position to the beginning.
-            json.dump(data, f)
-            f.truncate()# remove remaining part
-            
         
         return form_data, rejection_data
